+++ src/game/tests.rs
@@ -533,6 +533,7 @@ mod banding_blocked;
 mod banding_formation;
 mod banding_with_other;
 mod blaze_of_glory;
+mod blessing_and_flame;
 mod block_restriction_cards;
 mod blocking_after_death;
 mod blocking_prevention;